+++ app/components/Dashboard/risk_assesment/modify_rnc.js
@@ -1,12 +1,12 @@
+import { Icon } from 'native-base';
 import React, { Component } from 'react';
-import { View, Text, ScrollView, TouchableOpacity, ToastAndroid, Alert } from 'react-native';
-import { DataTable } from 'react-native-paper'
-import { rassessment_styles } from '../../../assets/styles/risk_assessment_styles'
-import { defaults } from '../../../assets/styles/default_styles'
-import { Icon } from 'native-base'
+import { Alert, ScrollView, Text, ToastAndroid, TouchableOpacity, View } from 'react-native';
+import { DataTable } from 'react-native-paper';
 import { NavigationEvents } from 'react-navigation';
-import Storage from '../../utils/storage'
-import Notification from '../../utils/alert_notification'
+import { defaults } from '../../../assets/styles/default_styles';
+import { rassessment_styles } from '../../../assets/styles/risk_assessment_styles';
+import Notification from '../../utils/alert_notification';
+import Storage from '../../utils/storage';
 
 export default class ModifyResourceAndCapacities extends Component {
   constructor(props) {
@@ -96,32 +96,34 @@ export default class ModifyResourceAndCapacities extends Component {
         let rnc_data = [];
         let to_local_data = [];
         let counter = 0
-        for (const [index, value] of responseJson.entries()) {
+        if (responseJson.length != 0) {
+          for (const [index, value] of responseJson.entries()) {
+            rnc_data.push(<DataTable.Row style={{ width: 500 }}>
+              <DataTable.Cell style={{ marginRight: 10 }}>{value.resource_and_capacity}</DataTable.Cell>
+              <DataTable.Cell style={{ marginRight: 10 }}>{value.status}</DataTable.Cell>
+              <DataTable.Cell style={{ marginRight: 10 }}>{value.owner}</DataTable.Cell>
+              <DataTable.Cell>
+                <Icon name="md-create" style={{ color: "blue" }} onPress={() => this.updateLog(value)}></Icon><Text>   </Text>
+                <Icon name="ios-trash" style={{ color: "red" }} onPress={() => this.removeConfirmation(value.resources_and_capacities_id)}></Icon>
+              </DataTable.Cell>
+            </DataTable.Row>)
+            counter += 1
+            to_local_data.push({
+              resources_and_capacities_id: value.resources_and_capacities_id,
+              local_storage_id: counter,
+              sync_status: 3,
+              resource_and_capacity: value.resource_and_capacity,
+              status: value.status,
+              owner: value.owner
+            })
+          }
+          Storage.removeItem("RiskAssessmentRNC")
+          Storage.setItem("RiskAssessmentRNC", to_local_data)
+        } else {
           rnc_data.push(<DataTable.Row style={{ width: 500 }}>
-            <DataTable.Cell style={{ marginRight: 10 }}>{value.resource_and_capacity}</DataTable.Cell>
-            <DataTable.Cell style={{ marginRight: 10 }}>{value.status}</DataTable.Cell>
-            <DataTable.Cell style={{ marginRight: 10 }}>{value.owner}</DataTable.Cell>
-            <DataTable.Cell>
-              <Icon name="md-create" style={{ color: "blue" }} onPress={() => this.updateLog(value)}></Icon><Text>   </Text>
-              <Icon name="ios-trash" style={{ color: "red" }} onPress={() => this.removeConfirmation(value.resources_and_capacities_id)}></Icon>
-            </DataTable.Cell>
+            <DataTable.Cell style={{ marginRight: 10 }}>No data</DataTable.Cell>
           </DataTable.Row>)
-          counter += 1
-          to_local_data.push({
-            resources_and_capacities_id: value.resources_and_capacities_id,
-            local_storage_id: counter,
-            sync_status: 3,
-            resource_and_capacity: value.resource_and_capacity,
-            status: value.status,
-            owner: value.owner
-          })
         }
-        Storage.removeItem("RiskAssessmentRNC")
-        Storage.setItem("RiskAssessmentRNC", to_local_data)
-        let data_container = Storage.getItem('RiskAssessmentRNC')
-        data_container.then(response => {
-          console.log(response)
-        });
         this.setState({ rnc_data: rnc_data })
         this.tablePaginate(rnc_data)
       })
